frontend-90: moved studio pages to pages folder, fixed namings
