+++ frontend/src/components/app/app.tsx
@@ -10,11 +10,10 @@ import { Search } from 'components/search/search';
 import { NotFound } from 'components/not-found-page/not-found';
 import { ReactNotifications } from 'react-notifications-component';
 import { ConfirmationModalTest } from './tests/confirmation-modal/confirmation-modal';
-import { StudioHome, StudioAnalytics, StudioSidebar } from '../studio';
+import { StudioHome, StudioAnalytics, StudioSidebar, StudioChannel } from '../../pages/studio';
 import { VideoCardTest } from './tests/video-card/video-card';
 import { VideoPageContainer } from 'pages/video/video-page-container';
 import { ProtectedRoute } from 'components/common/protected-route/protected-route';
-import { StudioChannel } from 'components/studio/channel/channel';
 import { AccountVerificationConfirmPage } from 'pages/account-verification-page/account-verification-confirm-page';
 import { RestorePasswordConfirmPage } from 'pages/restore-password-confirm-page/restore-password-confirm-page';
 import { ProfilePreferencesPage } from 'pages/profile-preferences-page/profile-preferences-page';
