+++ frontend/src/helpers/routes/is-router-has-default-navigation/is-route-has-default-navigation.helper.ts
