+++ frontend/src/pages/studio/analytics/styles.module.scss
