+++ frontend/src/pages/studio/channel/index.ts
@@ -1,0 +1,1 @@
+export { StudioChannel } from './channel';
