+++ frontend/src/pages/studio/common/sidebar/config.ts
