+++ frontend/src/pages/studio/common/sidebar/index.ts
