+++ frontend/src/pages/studio/common/sidebar/sidebar.tsx
