+++ frontend/src/pages/studio/common/sidebar/styles.module.scss
