+++ frontend/src/pages/studio/home/styles.module.scss
