+++ frontend/src/pages/studio/index.ts
@@ -1,3 +1,4 @@
 export { StudioHome } from './home';
 export { StudioSidebar } from './common/sidebar';
 export { StudioAnalytics } from './analytics';
+export { StudioChannel } from './channel';
